Refactor:리팩토링 카드 컴포넌트 분리
