+++ pages/shared/[id].tsx
@@ -3,7 +3,7 @@ import FolderUser from '@/src/components/folderUser/FolderUser';
 import Nav from '@/src/components/nav/Nav';
 import Footer from '@/src/components/footer/Footer';
 import SearchBar from '@/src/components/main/SearchBar';
-import CardList from '@/src/components/main/CardList';
+import SharedCard from '@/src/components/main/SharedCard';
 
 export default function SharedPage() {
   return (
@@ -18,7 +18,7 @@ export default function SharedPage() {
         </header>
         <main>
           <SearchBar />
-          <CardList />
+          <SharedCard />
         </main>
         <footer>
           <Footer />
