+++ src/components/main/Main.tsx
@@ -1,9 +1,9 @@
 import ButtonOption from '@/src/components/main/ButtonOption';
-import CardList from '@/src/components/main/CardList';
 import MainProvider from '@/src/components/main/MainProvider';
 import TabMenu from '@/src/components/main/TabMenu';
 import SearchBar from '@/src/components/main/SearchBar';
 import Modal from '@/src/components/modal/Modal';
+import FolderCard from './FolderCard';
 
 export default function Main() {
   return (
@@ -13,7 +13,7 @@ export default function Main() {
         <SearchBar />
         <TabMenu />
         <ButtonOption />
-        <CardList />
+        <FolderCard />
       </MainProvider>
     </>
   );
